+++ springboot/src/main/java/com/sqn/library/entity/Activity.java
@@ -3,12 +3,11 @@ package com.sqn.library.entity;
 import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
-
-import java.io.Serializable;
-
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
+
+import java.io.Serializable;
 
 /**
  * <p>
@@ -27,7 +26,7 @@ public class Activity implements Serializable {
 
     @ApiModelProperty("活动id")
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     @ApiModelProperty("活动名称")
     private String name;
